feat(module): 서버 별 빌드 파일 분리
