+++ package.json
@@ -6,10 +6,16 @@
   "private": true,
   "license": "UNLICENSED",
   "scripts": {
-    "build": "nest build",
+    "build": "MODE=APP nest build",
+    "build:was": "MODE=WAS nest build",
+    "build:game": "MODE=GAME nest build",
+    "build:media": "MODE=MEDIA nest build",
     "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
     "start": "nest start",
     "start:dev": "nest start --watch",
+    "start:dev:was": "MODE=WAS nest start --watch",
+    "start:dev:game": "MODE=GAME nest start --watch",
+    "start:dev:media": "MODE=MEDIA nest start --watch",
     "start:debug": "nest start --debug --watch",
     "start:prod": "node dist/src/main",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
