+++ src/app/app-was.module.ts
@@ -1,0 +1,28 @@
+import { Module } from '@nestjs/common';
+import { UsersModule } from '~/src/domain/users/users.module';
+import { ConfigModule, ConfigService } from '@nestjs/config';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { AuthModule } from '~/src/domain/auth/auth.module';
+import { AuctionModule } from '~/src/domain/auction/auction.module';
+import { FileModule } from '../domain/file/file.module';
+import { AwsModule } from '../domain/aws/aws.module';
+import { typeormConfig } from '~/src/app/config';
+
+@Module({
+  imports: [
+    ConfigModule.forRoot({
+      isGlobal: true,
+    }),
+    TypeOrmModule.forRootAsync({
+      imports: [ConfigModule],
+      useFactory: typeormConfig,
+      inject: [ConfigService],
+    }),
+    UsersModule,
+    AuthModule,
+    AuctionModule,
+    FileModule,
+    AwsModule,
+  ],
+})
+export class WasAppModule {}
