+++ src/main.ts
@@ -1,14 +1,35 @@
 import { NestFactory } from '@nestjs/core';
-import { AppModule } from './app.module';
+import { AppModule } from './app/app.module';
+import { GameAppModule } from '~/src/app/app-game.module';
+import { WasAppModule } from '~/src/app/app-was.module';
+import { MediaAppModule } from '~/src/app/app-media.module';
 import { ValidationPipe, VersioningType } from '@nestjs/common';
 import { setupSwagger } from '~/infrastructure/document/document.swagger';
 import { ServiceExceptionToHttpExceptionFilter } from '~/src/common/filters/exception-filter';
 import * as cookieParser from 'cookie-parser';
 
 async function bootstrap() {
-  const app = await NestFactory.create(AppModule);
+  const mode = process.env.MODE || 'APP';
+  let app = null;
+  switch (mode) {
+    case 'WAS':
+      console.log('Was App Start');
+      app = await NestFactory.create(WasAppModule);
+      break;
+    case 'GAME':
+      console.log('Game App Start');
+      app = await NestFactory.create(GameAppModule);
+      break;
+    case 'MEDIA':
+      console.log('Media App Start');
+      app = await NestFactory.create(MediaAppModule);
+      break;
+    default:
+      console.log('App Start');
+      app = await NestFactory.create(AppModule);
+      break;
+  }
   app.setGlobalPrefix('api');
-
   app.enableVersioning({
     type: VersioningType.URI,
     defaultVersion: '1',
